first implementation of the render
scalafrag code used. I started a basic html division of the desired frontend. I made the head that imports the correct font of the text, and the body that contains the players, the poker's table and the 3 buttons raise, check/call and fold. I started to create values to update the view with the correct current parameters.