CardSymbols method and css reviewed
Implementation reviewed in the types.scala file, to exclude the cavalier's unicode.
Tests created to verify the functionnality of the method in TestTypes.scala.
I made the render method in the UI.scala file better by grouping the code and
made it generate players from 0 to 5 depending on the numbers of clients.
I added a case frag when it's the end of turn as a sketch frag.