Comments added
I started to comment the UI.scala file, and I removed the unnecessary calls of the userId in all the render's functions. I also added a text to be displayed while a player waits for the others. I used the name of the current player to display the last action that happened.